replace pcoin edit table 1, correct filter traits
